requirements.txt
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
blinker==1.6.2
Bottleneck==1.3.5
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.2
filelock==3.9.0
fonttools==4.43.1
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.0
mkl-random==1.2.4
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio==1.5.6
numba==0.57.1
numexpr==2.8.7
numpy==1.24.3
orjson==3.9.9
oscrypto==1.2.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
plotly==5.17.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==2.0.3
pydantic_core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydeck==0.8.0b4
Pygments==2.16.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
setuptools==68.2.2
shap==0.43.0
sip==6.6.2
six==1.16.0
slicer==0.0.7
smmap==3.0.5
sniffio==1.3.0
snowflake-connector-python==3.2.0
sortedcontainers==2.4.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.27.2
streamlit-searchbox==0.1.5
tenacity==8.2.3
threadpoolctl==2.2.0
toml==0.10.2
tomlkit==0.11.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.8
websockets==11.0.3
wheel==0.41.2
widgetsnbextension==4.0.9
win-inet-pton==1.1.0
zipp==3.17.0
